# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RoundApp OpenApi Documentation",
        "description": "L5 Swagger OpenApi",
        "contact": {
            "email": "francesco.simeoli@gmail.com"
        },
        "license": {
            "name": "Apache 2.0 License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://roundapp.local/api",
            "description": "RoundApp API Server"
        },
        {
            "url": "https://roundapp.local"
        }
    ],
    "paths": {
...
        "/auth/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Password reset",
                "description": "Password reset",
                "operationId": "authPasswordReset",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "token": "ce31036da62de3d297ab806cfa72c6d3a9d7215118463346a83a4d1cb5cc9c37",
                                    "email": "test@roundapp.local",
                                    "password": "12345678",
                                    "password_confirmation": "12345678"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (invalid token, invalid email, invalid password)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "message": "The given data was invalid.",
                                        "errors": {
                                            "email": "We can't find a user with that email address."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://roundapp.local/oauth/authorize",
                        "tokenUrl": "https://roundapp.local/oauth/token",
                        "refreshUrl": "https://roundapp.local/token/refresh",
                        "scopes": [],
                        "client_id": "1",
                        "secret": "yo4kJNkiGGIiU0YqZ5jP1ucCArLM11FC7z5QdEFY"
                    }
                }
            }
        }
    },
...
}